requirements_mcp.txt
# Secure MCP Gateway - Dependencies
mcp>=1.0.0
uvicorn>=0.30.0
starlette>=0.38.0
httpx>=0.27.0

# Sécurité — chiffrement des tokens au repos :
cryptography>=43.0.0

# OAuth2 multi-tenant — JWT pour les sessions Claude Desktop :
PyJWT>=2.8.0

# Gmail / Google Calendar :
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Slack :
slack-sdk>=3.30.0

# HubSpot (décommenter pour la production) :
# hubspot-api-client>=9.0.0

# GitHub (décommenter pour la production) :
# PyGithub>=2.0.0

# --- Persistance (Option C) ---

# PostgreSQL (async) :
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0

# Redis (async) :
redis[hiredis]>=5.0.0

# Migrations DB :
alembic>=1.13.0
